# api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenSlide StarDist Viewer API
  version: 0.1.0
  description: Tile streaming, nuclei segmentation, and cell counting for local pathology slides.
servers:
  - url: http://localhost:25342
paths:
  /healthz:
    get:
      operationId: getHealthz
      responses:
        "200":
          description: Health status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /readyz:
    get:
      operationId: getReadyz
      responses:
        "200":
          description: Readiness status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
  /api/slides:
    get:
      operationId: listSlides
      responses:
        "200":
          description: Available slides.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlideListResponse"
  /api/slides/{slide_id}:
    get:
      operationId: getSlide
      parameters:
        - $ref: "#/components/parameters/SlideId"
      responses:
        "200":
          description: Slide metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlideMetadata"
        "404":
          $ref: "#/components/responses/Error"
  /api/slides/{slide_id}/dzi:
    get:
      operationId: getSlideDzi
      parameters:
        - $ref: "#/components/parameters/SlideId"
      responses:
        "200":
          description: Deep Zoom XML descriptor.
          content:
            application/xml:
              schema:
                type: string
  /api/slides/{slide_id}_files/{level}/{column}_{row}.jpeg:
    get:
      operationId: getSlideTile
      parameters:
        - $ref: "#/components/parameters/SlideId"
        - name: level
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
        - name: column
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
        - name: row
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: JPEG tile.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
  /api/slides/{slide_id}/thumbnail:
    get:
      operationId: getSlideThumbnail
      parameters:
        - $ref: "#/components/parameters/SlideId"
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 128
            maximum: 2048
            default: 900
      responses:
        "200":
          description: JPEG thumbnail.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
  /api/slides/{slide_id}/segment:
    post:
      operationId: segmentSlideRegion
      parameters:
        - $ref: "#/components/parameters/SlideId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SegmentRequest"
      responses:
        "200":
          description: Nuclei segmentation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentResponse"
        "413":
          $ref: "#/components/responses/Error"
components:
  parameters:
    SlideId:
      name: slide_id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: JSON API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ReadyResponse:
      type: object
      required: [status, slides]
      properties:
        status:
          type: string
        slides:
          type: integer
    Dimensions:
      type: object
      required: [width, height]
      properties:
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
    SlideMetadata:
      type: object
      required: [id, name, filename, format, dimensions, level_count, level_dimensions, tile_size, properties]
      properties:
        id:
          type: string
        name:
          type: string
        filename:
          type: string
        format:
          type: string
        dimensions:
          $ref: "#/components/schemas/Dimensions"
        level_count:
          type: integer
        level_dimensions:
          type: array
          items:
            $ref: "#/components/schemas/Dimensions"
        tile_size:
          type: integer
        mpp_x:
          type: number
          nullable: true
        mpp_y:
          type: number
          nullable: true
        objective_power:
          type: number
          nullable: true
        properties:
          type: object
          additionalProperties:
            type: string
    SlideListResponse:
      type: object
      required: [slides]
      properties:
        slides:
          type: array
          items:
            $ref: "#/components/schemas/SlideMetadata"
    SegmentRequest:
      type: object
      required: [x, y, width, height]
      properties:
        x:
          type: integer
          minimum: 0
        y:
          type: integer
          minimum: 0
        width:
          type: integer
          minimum: 1
          maximum: 4096
        height:
          type: integer
          minimum: 1
          maximum: 4096
        max_nuclei:
          type: integer
          minimum: 1
          maximum: 10000
          default: 2500
    Region:
      type: object
      required: [x, y, width, height]
      properties:
        x:
          type: integer
        y:
          type: integer
        width:
          type: integer
        height:
          type: integer
    Nucleus:
      type: object
      required: [id, centroid, area, bbox, polygon]
      properties:
        id:
          type: integer
        centroid:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
        area:
          type: number
        bbox:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: number
        confidence:
          type: number
          nullable: true
        polygon:
          type: array
          items:
            type: array
            minItems: 2
            maxItems: 2
            items:
              type: number
    SegmentResponse:
      type: object
      required: [slide_id, method, region, count, elapsed_ms, nuclei]
      properties:
        slide_id:
          type: string
        method:
          type: string
        region:
          $ref: "#/components/schemas/Region"
        count:
          type: integer
        elapsed_ms:
          type: number
        nuclei:
          type: array
          items:
            $ref: "#/components/schemas/Nucleus"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true